Split TrailblockAgent into its own file.
It is a bit big and complex to be lumped into one file with other stuff.
